+++ AmongUsCaptureConsole/Properties/PublishProfiles/FolderProfile.pubxml
