+++ AmongUsCaptureUI/App.xaml
@@ -1,4 +1,4 @@
-﻿<Application x:Class="AUCapture_WPF.App"
+﻿<Application x:Class="AmongUsCaptureUI.App"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml">
     <Application.Resources>
